Refactor Env class in testsuite.
- read nodes and chains environments form file and apply it.
- add Channel class to Env.

+++ testsuite/java/foundation/icon/test/score/SampleTokenScore.java
@@ -74,7 +74,7 @@ public class SampleTokenScore extends Score {
                 .build();
 
         Transaction transaction = TransactionBuilder.newBuilder()
-                .nid(chain.networkId)
+                .nid(BigInteger.valueOf(chain.networkId))
                 .from(fromWallet.getAddress())
                 .to(scoreAddress)
                 .stepLimit(STEPS_DEFAULT)
